pi-aggregator/requirements.txt
paho-mqtt
numpy
pandas
python-dateutil
